+++ source/renderer/dialog/preferences.js
@@ -163,7 +163,7 @@ class PreferencesDialog extends ZettlrDialog {
     if (cfg.hasOwnProperty('attachmentExtensions')) {
       let attachments = cfg['attachmentExtensions'].split(',')
       for (let i = 0; i < attachments.length; i++) {
-        attachments[i] = attachments[i].trim().replace(/[\s]/g, '')
+        attachments[i] = attachments[i].trim().replace(/\s/g, '')
         if (attachments[i].length < 2) {
           attachments.splice(i, 1)
           i--
@@ -172,6 +172,9 @@ class PreferencesDialog extends ZettlrDialog {
         if (attachments[i].charAt(0) !== '.') {
           attachments[i] = '.' + attachments[i]
         }
+
+        // Convert to lower case
+        attachments[i] = attachments[i].toLowerCase()
       }
       cfg['attachmentExtensions'] = attachments
     }